+++ sdl3-ttf-sys/ChangeLog.md
@@ -1,8 +1,8 @@
 # Changes
 
 - next:
-    - Add `from_raw` assoc function and `into_raw` and `value` methods for groups (ids, enums, etc),
-      and impl Display for ids under the `display-impls` feature
+    - Add `new` and `value` assoc function/method for groups (ids, enums, etc)
+    - impl Display for ids under the `display-impls` feature
     - Add metadata for structs and unions
 
 - 0.6.0: